[REF] emc_connector: split adapters into seperate files
