+++ easy_my_coop_connector/components/__init__.py
@@ -1,3 +1,2 @@
-from . import emc_adapters
 from . import emc_backend
 from . import emc_bindings
