+++ easy_my_coop_connector/components/subscription_request_adapter.py
@@ -2,66 +2,13 @@
 #   Robin Keunen <robin@coopiteasy.be>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from urllib.parse import urljoin
-
 from werkzeug.exceptions import BadRequest
 
 from odoo import _
 from odoo.exceptions import UserError, ValidationError
 from odoo.fields import Date
 
-
-class AbstractEMCAdapter:
-    _model = "set in implementation class"
-    _root = "/api/"
-    _service = "set in implementation class"
-
-    def __init__(self, backend, record=None):
-        self.backend = backend
-        self.record = record
-
-    def _get_url(self, args=None):
-        """args is a list of path elements
-        :return the complete route to the service
-        """
-        if args is None:
-            args = []
-        return urljoin(self._root, self._service, *args)
-
-    def search(self, **params):
-        raise NotImplementedError
-
-    def read(self, id_):
-        # pylint: disable=method-required-super
-        url = self._get_url([str(id_)])
-        api_dict = self.backend.http_get_content(url)
-        return self.to_write_values(api_dict)
-
-    def create(self, record):
-        # pylint: disable=method-required-super
-        url = self._get_url()
-        api_dict = self.to_api_dict(record)
-        external_record = self.backend.http_post_content(url, api_dict)
-        external_id, writeable_dict = self.to_write_values(external_record)
-        return external_id, writeable_dict
-
-    def update(self, data):
-        raise NotImplementedError
-
-    def delete(self):
-        raise NotImplementedError
-
-    def to_write_values(self, api_dict):
-        """
-        :return a tuple with
-            - the external id
-            - a writable dictionary for _model
-        received from the api
-        """
-        raise NotImplementedError
-
-    def to_api_dict(self, record):
-        raise NotImplementedError
+from .abstract_emc_adapter import AbstractEMCAdapter
 
 
 class SubscriptionRequestAdapter(AbstractEMCAdapter):
@@ -191,54 +138,3 @@ class SubscriptionRequestAdapter(AbstractEMCAdapter):
             "capital_release_request": crr_tuples,
         }
         return external_id, writable_dict
-
-
-class AccountInvoiceAdapter(AbstractEMCAdapter):
-    _model = "account.invoice"
-    _service = "invoice"
-
-    def to_write_values(self, api_dict):
-        external_id = api_dict.pop("id")
-        writable_dict = api_dict
-        return external_id, writable_dict
-
-
-class AccountPaymentAdapter(AbstractEMCAdapter):
-    _model = "account.payment"
-    _service = "payment"
-
-    def to_write_values(self, api_dict):
-        api_dict = api_dict.copy()
-        external_id = api_dict.pop("id")
-        writable_dict = api_dict
-        return external_id, writable_dict
-
-    def to_api_dict(self, record):
-
-        if not record.journal_id.binding_id:
-            raise ValidationError(
-                _(
-                    "Journal %s is not bound to a journal on the platform. "
-                    "Please contact system administrator."
-                )
-                % record.journal_id.name
-            )
-
-        if not record.invoice_ids.binding_id:
-            raise ValidationError(
-                _(
-                    "Invoice %s is not bound to a journal on the platform. "
-                    "Please contact system administrator."
-                )
-                % record.invoice_ids.name
-            )
-
-        return {
-            "journal": record.journal_id.binding_id.external_id,
-            "invoice": record.invoice_ids.binding_id.external_id,
-            "payment_date": Date.to_string(record.payment_date),
-            "amount": record.amount,
-            "communication": record.communication,
-            "payment_type": record.payment_type,
-            "payment_method": record.payment_method_id.code,
-        }
