+++ easy_my_coop_connector/models/subscription_request.py
@@ -7,7 +7,9 @@ from datetime import date, timedelta
 
 from odoo import api, fields, models
 
-from ..components.emc_adapters import SubscriptionRequestAdapter
+from ..components.subscription_request_adapter import (
+    SubscriptionRequestAdapter,
+)
 
 _logger = logging.getLogger(__name__)
